refactor: move the header and the media view
